refactor: Move runScript to it's own file.
